move genesis to its own module
