+++ crates/astria-sequencer/src/accounts/component.rs
@@ -9,11 +9,8 @@ use tracing::instrument;
 
 use super::state_ext::StateWriteExt;
 use crate::{
-    app::{
-        Account,
-        GenesisState,
-    },
     component::Component,
+    genesis::GenesisState,
 };
 
 #[derive(Default)]
@@ -25,12 +22,8 @@ impl Component for AccountsComponent {
 
     #[instrument(name = "AccountsComponent:init_chain", skip(state))]
     async fn init_chain<S: StateWriteExt>(mut state: S, app_state: &Self::AppState) -> Result<()> {
-        for Account {
-            address,
-            balance,
-        } in &app_state.accounts
-        {
-            state.put_account_balance(address, *balance)?;
+        for account in &app_state.accounts {
+            state.put_account_balance(&account.address, account.balance)?;
         }
         Ok(())
     }
